cleanup custom skin color codes feature
