+++ src/actors/mario/model.inc.js
@@ -15,99 +15,13 @@ export const mario_texture_hair_sideburn = []
 export const mario_texture_mustache = []
 
 // 0x04000000 # solid color blue - butt, left thigh, right thigh - all poly types
-/*const mario_overalls_lights_group = [
-	Gbi.gdSPDefLights1(
-		0x00, 0x00, 0x7f,
-		0x00, 0x00, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x00, 0x7f,
-		0xff, 0x00, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x7f, 0x7f,
-		0xff, 0xff, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x60, 0x3c,
-		0xfe, 0xc1, 0x79, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x00, 0x00,
-		0xff, 0x00, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x00, 0x7f,
-		0xff, 0x00, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x39, 0x0e, 0x07,
-		0x72, 0x1c, 0x0e, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x00, 0x00,
-		0xff, 0x00, 0x00, 0x28, 0x28, 0x28
-	),
-]
-*/
 const mario_overalls_lights_group = (customData) => {
-	return Gbi.gdSPDefLights1(...customData.mario_overalls_lights)
+	return Gbi.gdSPDefLights1(...customData.mario_overalls_lights, 0x28, 0x28, 0x28)
 }
 
 // 0x04000018 # solid color red - left & right arm, torso (tshirt part), caps - all poly types
-/*const mario_hat_shirt_lights_group = [
-	Gbi.gdSPDefLights1(
-		0x7f, 0x00, 0x00,
-		0xff, 0x00, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x7f, 0x00,
-		0xff, 0xff, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x7f, 0x00,
-		0xff, 0xff, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x7f, 0x7f,
-		0xff, 0xff, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x39, 0x0e, 0x07,
-		0x72, 0x1c, 0x0e, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x00, 0x00, 0x7f,
-		0x00, 0x00, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x00, 0x7f, 0x7f,
-		0x00, 0xff, 0xff, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x00, 0x7f, 0x00,
-		0x00, 0xff, 0x00, 0x28, 0x28, 0x28
-	),
-	Gbi.gdSPDefLights1(
-		0x7f, 0x7f, 0x00,
-		0xff, 0xff, 0x00, 0x28, 0x28, 0x28
-	),
-]
-*/
 const mario_hat_shirt_lights_group = (customData) => {
-	return Gbi.gdSPDefLights1(...customData.mario_hat_shirt_lights)
+	return Gbi.gdSPDefLights1(...customData.mario_hat_shirt_lights, 0x28, 0x28, 0x28)
 }
 
 // 0x04000030 # solid color white - metal butt & left thigh - normal left & right hand closed & open (with cap too) and all wings - all poly types
