+++ src/cosmetics.js
@@ -1,0 +1,39 @@
+import { networkData } from "./socket"
+
+const overallsPresets = [
+    [0x00, 0x00, 0x7f, 0x00, 0x00, 0xff, 0x28, 0x28, 0x28],
+    [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x28, 0x28],
+    [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x28, 0x28],
+    [0x7f, 0x00, 0x7f, 0xff, 0x00, 0xff, 0x28, 0x28, 0x28],
+    [0x7f, 0x7f, 0x7f, 0xff, 0xff, 0xff, 0x28, 0x28, 0x28],
+    [0x7f, 0x60, 0x3c, 0xfe, 0xc1, 0x79, 0x28, 0x28, 0x28],
+    [0x7f, 0x00, 0x00, 0xff, 0x00, 0x00, 0x28, 0x28, 0x28],
+    [0x7f, 0x00, 0x7f, 0xff, 0x00, 0xff, 0x28, 0x28, 0x28],
+    [0x39, 0x0e, 0x07, 0x72, 0x1c, 0x0e, 0x28, 0x28, 0x28],
+    [0x7f, 0x00, 0x00, 0xff, 0x00, 0x00, 0x28, 0x28, 0x28]
+]
+
+const hatShirtPresets = [
+    [ 0x7f, 0x00, 0x00, 0xff, 0x00, 0x00, 0x28, 0x28, 0x28 ],
+    [ 0x7f, 0x7f, 0x00, 0xff, 0xff, 0x00, 0x28, 0x28, 0x28 ],
+    [ 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x28, 0x28 ],
+    [ 0x7f, 0x7f, 0x00, 0xff, 0xff, 0x00, 0x28, 0x28, 0x28 ],
+    [ 0x7f, 0x7f, 0x7f, 0xff, 0xff, 0xff, 0x28, 0x28, 0x28 ],
+    [ 0x39, 0x0e, 0x07, 0x72, 0x1c, 0x0e, 0x28, 0x28, 0x28 ],
+    [ 0x00, 0x00, 0x7f, 0x00, 0x00, 0xff, 0x28, 0x28, 0x28 ],
+    [ 0x00, 0x7f, 0x7f, 0x00, 0xff, 0xff, 0x28, 0x28, 0x28 ],
+    [ 0x00, 0x7f, 0x00, 0x00, 0xff, 0x00, 0x28, 0x28, 0x28 ],
+    [ 0x7f, 0x7f, 0x00, 0xff, 0xff, 0x00, 0x28, 0x28, 0x28 ]
+]
+
+window.updateSkinID = (skinID) => {
+    window.myMario.overalls = overallsPresets[skinID]
+    window.myMario.hatShirt = hatShirtPresets[skinID]
+}
+
+export const recvSkinData = (msg) => { 
+    if (msg.channel_id != networkData.mySocketID &&
+        networkData.remotePlayers[msg.channel_id] == undefined) return
+
+    networkData.remotePlayers[msg.channel_id].skinData = msg.skinData
+}
